+++ src/app/pages/dashboard/dashboard.component.ts
@@ -1,15 +1,10 @@
 import { BreakpointObserver, Breakpoints } from '@angular/cdk/layout'
 import { Component } from '@angular/core'
+import { AngularFirestore } from '@angular/fire/firestore'
+import { Chance } from 'chance'
 import { Observable } from 'rxjs'
 import { map } from 'rxjs/operators'
 import { DashboardService } from './dashboard.service'
-
-export interface DashboardMenuContent {
-  icon: string
-  title: string
-  subTitle: string
-  route: string
-}
 
 @Component({
   selector: 'app-dashboard',
@@ -17,45 +12,6 @@ export interface DashboardMenuContent {
   styleUrls: ['./dashboard.component.scss']
 })
 export class DashboardComponent {
-  cardsContent: DashboardMenuContent[] = [
-    {
-      icon: 'insert_chart',
-      title: 'LEAD Points',
-      subTitle: 'Add LEAD points to one user',
-      route: 'points'
-    },
-    {
-      icon: 'group_add',
-      title: 'Bulk Edit',
-      subTitle: 'Add points to multiple users at once',
-      route: 'bulk'
-    },
-    {
-      icon: 'supervised_user_circle',
-      title: 'User Management',
-      subTitle: 'A list with details of all users',
-      route: 'user-management'
-    },
-    {
-      icon: 'assignment',
-      title: 'Logs',
-      subTitle: 'Check all the changes that have been made',
-      route: 'logs'
-    },
-    {
-      icon: 'assignment',
-      title: 'Logs Management',
-      subTitle: 'Check all the changes that have been made',
-      route: 'logs-management'
-    },
-    {
-      icon: 'delete_forever',
-      title: 'Reset All Points',
-      subTitle: 'Reset the LEAD Points Leaderboard',
-      route: 'reset'
-    }
-  ]
-
   isHandset$: Observable<boolean> = this.breakpointObserver
     .observe(Breakpoints.Handset)
     .pipe(map(result => result.matches))
@@ -63,12 +19,41 @@ export class DashboardComponent {
   /** Based on the screen size, switch from standard to one column per row */
   cards = this.breakpointObserver.observe(Breakpoints.Handset).pipe(
     map(() => {
-      return [...this.cardsContent]
+      return [...this.dashboardService.navItems]
     })
   )
 
   constructor(
     public dashboardService: DashboardService,
-    private breakpointObserver: BreakpointObserver
-  ) {}
+    private breakpointObserver: BreakpointObserver,
+    private afs: AngularFirestore
+  ) {
+    // this.generateUsers()
+  }
+
+  generateUsers() {
+    const campus = ['Sydney', 'Melbourne']
+
+    setInterval(() => {
+      const chance = new Chance()
+      const user = {
+        uid: chance.string({
+          length: 28,
+          pool: 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789'
+        }),
+        photoURL: chance.avatar({ protocol: 'https' }),
+        displayName: chance.name({ middle: true }),
+        email: `k123456@student.kent.edu.au`,
+        campus: campus[chance.integer({ min: 0, max: 1 })],
+        incognitoMode: chance.bool(),
+        role: 'user',
+        points: chance.integer({ min: 0, max: 1000 })
+      }
+
+      this.afs
+        .collection('users')
+        .doc(user.uid)
+        .set(user)
+    }, 200)
+  }
 }