+++ src/app/pages/dashboard/dashboard.service.ts
@@ -1,13 +1,11 @@
 import { TitleCasePipe } from '@angular/common'
-import { Injectable } from '@angular/core'
-import { AngularFirestoreDocument } from '@angular/fire/firestore'
+import { Injectable, ViewChild } from '@angular/core'
 import { MatDialog } from '@angular/material/dialog'
+import { MatSidenav } from '@angular/material/sidenav'
 import { Router } from '@angular/router'
 import { ResetPointsComponent } from '@dialogs/reset-points/reset-points.component'
-import {
-  // Log,
-  NewLog
-} from '@interfaces/log'
+import { Log } from '@interfaces/log'
+import { NavItem } from '@interfaces/nav-item'
 import { User } from '@interfaces/user'
 // import { AuthService } from '@services/auth.service'
 import { FirestoreService } from '@services/firestore.service'
@@ -17,6 +15,53 @@ import { ToastService } from '@services/toast.service'
   providedIn: 'root'
 })
 export class DashboardService {
+  @ViewChild('sidenav') sidenav: MatSidenav
+
+  navItems: NavItem[] = [
+    {
+      icon: 'home',
+      title: 'Dashboard',
+      subTitle: 'Dashboard',
+      routePath: 'dashboard'
+    },
+    {
+      icon: 'insert_chart',
+      title: 'Add Points',
+      subTitle: 'Add LEAD points to one user',
+      routePath: 'points'
+    },
+    // {
+    //   icon: 'group_add',
+    //   title: 'Add Bulk Points',
+    //   subTitle: 'Add points to multiple users at once',
+    //   routePath: 'bulk'
+    // },
+    // {
+    //   icon: 'supervised_user_circle',
+    //   title: 'User Management',
+    //   subTitle: 'A list with details of all users',
+    //   routePath: 'user-management'
+    // },
+    {
+      icon: 'assignment',
+      title: 'Logs',
+      subTitle: 'Check all the changes that have been made',
+      routePath: 'logs'
+    },
+    // {
+    //   icon: 'assignment',
+    //   title: 'Logs Management',
+    //   subTitle: 'Check all the changes that have been made',
+    //   routePath: 'logs-management'
+    // },
+    {
+      icon: 'delete_forever',
+      title: 'Reset Points',
+      subTitle: 'Reset the LEAD Points Leaderboard',
+      routePath: 'reset'
+    }
+  ]
+
   constructor(
     private toast: ToastService,
     private fss: FirestoreService,
@@ -41,37 +86,56 @@ export class DashboardService {
   }
 
   async logData(points: number, pointsAdded: number, userRef: User) {
-    const displayName = this.titlecasePipe.transform(userRef.displayName)
+    const userName = this.titlecasePipe.transform(userRef.displayName)
 
-    const dataObj: NewLog = {
+    const dataObj: Log = {
       date: Date.now(),
-      message: `${displayName} now has ${points} pts`,
+      message: `${userName} now has ${points} pts`,
       pointsAdded,
       pointsCurrent: points,
-      userId: userRef.uid
+      userId: userRef.uid,
+      userName
     }
 
     this.fss
       .addLog(dataObj)
       .then(() => {
-        this.toast.showSuccess(
-          `You've ${this.logText(pointsAdded, displayName)}`
-        )
+        this.toast.showSuccess(`You've ${this.logText(pointsAdded, userName)}`)
       })
       .catch(error => {
         this.toast.showError(error)
       })
   }
 
-  async updateData(userRef: AngularFirestoreDocument<User>, data: User) {
-    return userRef.update(data)
-  }
+  //   async logData(points: number, pointsAdded: number, userRef: User) {
+  //   const displayName = this.titlecasePipe.transform(userRef.displayName)
+  //   return this.auth.user$.subscribe(admin => {
+  //     const dataObj: Log = {
+  //       log: `${displayName} now has ${points} pts`,
+  //       adminName: admin.displayName,
+  //       pointsAdded,
+  //       userName: displayName,
+  //       date: new Date().getTime()
+  //     }
+  //
+  //     this.fss
+  //       .addLog(dataObj)
+  //       .then(() => {
+  //         this.toast.showSuccess(
+  //           `You've ${this.logText(pointsAdded, displayName)}`
+  //         )
+  //       })
+  //       .catch(error => {
+  //         this.toast.showError(error)
+  //       })
+  //   })
+  // }
 
-  checkRoute(item: { route: string }) {
-    switch (item.route) {
+  checkRoute(item: { routePath: string }) {
+    switch (item.routePath) {
       case 'dashboard':
       case 'profile': {
-        this.router.navigate([`${item.route}`])
+        this.router.navigate([`${item.routePath}`])
         break
       }
       case 'reset': {
@@ -79,7 +143,7 @@ export class DashboardService {
         break
       }
       default: {
-        this.router.navigate([`/dashboard/${item.route}`])
+        this.router.navigate([`/dashboard/${item.routePath}`])
         break
       }
     }