+++ src/app/pages/dashboard/nav/nav.component.ts
@@ -1,16 +1,11 @@
 import { BreakpointObserver, Breakpoints } from '@angular/cdk/layout'
-import { Component } from '@angular/core'
-import { ActivatedRoute } from '@angular/router'
+import { Component, ViewChild } from '@angular/core'
+import { MatSidenav } from '@angular/material/sidenav'
+import { ActivatedRoute, NavigationEnd, Router } from '@angular/router'
+import { NavItem } from '@interfaces/nav-item'
 import { Observable } from 'rxjs'
-import { map } from 'rxjs/operators'
+import { filter, map } from 'rxjs/operators'
 import { DashboardService } from '../dashboard.service'
-
-export interface DashboardMenuContent {
-  icon: string
-  title: string
-  subTitle: string
-  route: string
-}
 
 @Component({
   selector: 'app-nav',
@@ -18,64 +13,39 @@ export interface DashboardMenuContent {
   styleUrls: ['./nav.component.scss']
 })
 export class NavComponent {
-  routeTitle: string
+  @ViewChild('sidenav') sidenav: MatSidenav
 
+  routeTitle: string
   isHandset$: Observable<boolean> = this.breakpointObserver
     .observe(Breakpoints.Handset)
     .pipe(map(result => result.matches))
 
-  cardsContent: DashboardMenuContent[] = [
-    {
-      icon: 'home',
-      title: 'Dashboard',
-      subTitle: 'Dashboard',
-      route: 'dashboard'
-    },
-    {
-      icon: 'insert_chart',
-      title: 'LEAD Points',
-      subTitle: 'Add LEAD points to one user',
-      route: 'points'
-    },
-    {
-      icon: 'group_add',
-      title: 'Bulk Edit',
-      subTitle: 'Add points to multiple users at once',
-      route: 'bulk'
-    },
-    {
-      icon: 'supervised_user_circle',
-      title: 'User Management',
-      subTitle: 'A list with details of all users',
-      route: 'user-management'
-    },
-    {
-      icon: 'assignment',
-      title: 'Logs',
-      subTitle: 'Check all the changes that have been made',
-      route: 'logs'
-    },
-    {
-      icon: 'assignment',
-      title: 'Logs Management',
-      subTitle: 'Check all the changes that have been made',
-      route: 'logs-management'
-    },
-    {
-      icon: 'delete_forever',
-      title: 'Reset All Points',
-      subTitle: 'Reset the LEAD Points Leaderboard',
-      route: 'reset'
-    }
-  ]
+  navItems: NavItem[]
 
   constructor(
     public dashboardService: DashboardService,
     private route: ActivatedRoute,
+    private router: Router,
     private breakpointObserver: BreakpointObserver
   ) {
-    this.route.data.subscribe(data => {
-      this.routeTitle = data.title
+    this.navItems = this.dashboardService.navItems
+
+    this.router.events
+      .pipe(filter(event => event instanceof NavigationEnd))
+      .subscribe(() => {
+        this.route.children[0].data.subscribe(data => {
+          this.routeTitle = data.title
+        })
+      })
+  }
+
+  redirectTo(routeParam: string) {
+    this.dashboardService.checkRoute({ routePath: routeParam })
+
+    this.isHandset$.subscribe(isMobile => {
+      if (isMobile) {
+        this.sidenav.close()
+      }
     })
   }
 }