+++ src/app/pages/dashboard/points/points.component.ts
@@ -1,17 +1,12 @@
 import { Component } from '@angular/core'
-import {
-  AngularFirestore,
-  AngularFirestoreDocument
-} from '@angular/fire/firestore'
-import { User } from '@interfaces/user'
-import { FirestoreService } from '@services/firestore.service'
-import { take } from 'rxjs/operators'
-import { DashboardService } from '../dashboard.service'
-
-// import { DialogConfirmationComponent } from '@dialogs/dialog-confirmation/dialog-confirmation.component'
-// import { MatDialog, MatDialogRef } from '@angular/material/dialog'
 import { FormControl } from '@angular/forms'
-// import { ToastService } from '@services/toast.service'
+import { MatDialog, MatDialogRef } from '@angular/material/dialog'
+import { ConfirmPointsComponent } from '@dialogs/confirm-points/confirm-points.component'
+import { User } from '@interfaces/user'
+import { DbService } from '@services/db.service'
+import { FirestoreService } from '@services/firestore.service'
+import { ToastService } from '@services/toast.service'
+import { DashboardService } from '../dashboard.service'
 
 @Component({
   selector: 'app-points',
@@ -19,52 +14,53 @@ import { FormControl } from '@angular/forms'
   styleUrls: ['./points.component.scss']
 })
 export class PointsComponent {
-  // dialogRef: MatDialogRef<DialogConfirmationComponent>
-  pointsInputForm = new FormControl('')
+  dialogRef: MatDialogRef<ConfirmPointsComponent>
+  pointsFormControl = new FormControl(null)
 
   constructor(
     public fss: FirestoreService,
-    // public dialog: MatDialog,
-    // private toast: ToastService,
+    private dialog: MatDialog,
+    private toast: ToastService,
     private dashboardService: DashboardService,
-    private afs: AngularFirestore
+    private db: DbService
   ) {}
 
-  // Add Points to one user onChange - Update the user value at the DB
-  update(user: User, newPoints: number) {
-    const userDoc: AngularFirestoreDocument<User> = this.afs.doc(
-      `users/${user.uid}`
-    )
+  checkPoints(user: User) {
+    const originalPoints = user.points
+    const newPoints: number = this.pointsFormControl.value
+    const limitPoints = 200
+    const addedPoints = newPoints - originalPoints
 
-    userDoc
-      .valueChanges()
-      .pipe(take(1))
-      .subscribe(userRef => {
-        const addedPoints = newPoints - userRef.points
-        const points = newPoints
+    if (addedPoints > limitPoints) {
+      this.openDialog()
+      this.dialogRef.afterClosed().subscribe(acceptedChanges => {
+        if (acceptedChanges) {
+          this.updatePoints(user, newPoints, addedPoints)
+        } else {
+          // TODO: reset points to it's original state
+        }
+      })
+    } else {
+      this.updatePoints(user, newPoints, addedPoints)
+    }
+  }
 
-        // <<<<<<< HEAD
-        //     userData$.subscribe(userRef => {
-        //       const addedPoints = newPoints - userRef.points
-        //       const points = Number(newPoints)
-        //
-        //       if (addedPoints > 700) {
-        //         return this.toast.showError(`Please add less than 700 points`)
-        //       }
-        // =======
-        //         const newData: User = {
-        //           points
-        //         }
-        // >>>>>>> feature-log-filter
-
-        // this.dashboardService.updateData(userDoc, newData)
-        this.dashboardService.logData(points, addedPoints, userRef)
+  private updatePoints(user: User, newPoints: number, addedPoints: number) {
+    this.db
+      .updateAt(`users/${user.uid}`, { points: Number(newPoints) })
+      .then(() => {
+        if (addedPoints) {
+          this.dashboardService.logData(newPoints, addedPoints, user)
+        }
+      })
+      .catch(err => {
+        this.toast.showError(err)
       })
   }
 
-  // openDialogConfirmation(): void {
-  //   this.dialogRef = this.dialog.open(DialogConfirmationComponent, {
-  //     autoFocus: false
-  //   })
-  // }
+  private openDialog(): void {
+    this.dialogRef = this.dialog.open(ConfirmPointsComponent, {
+      autoFocus: false
+    })
+  }
 }