+++ src/app/pages/dashboard/user-management/user-management.component.ts
@@ -1,9 +1,10 @@
 import { Component, OnInit, ViewChild } from '@angular/core'
-import { MatPaginator, MatSort, MatTableDataSource } from '@angular/material'
-import { User } from '@interfaces/user'
-import { UserData } from '@interfaces/user-data'
+import { MatPaginator, MatSort } from '@angular/material'
+// import { User } from '@interfaces/user'
+// import { UserData } from '@interfaces/user-data'
 import { FirestoreService } from '@services/firestore.service'
-import { map } from 'rxjs/operators'
+// import { map } from 'rxjs/operators'
+import { UserManagementDataSource } from './user-management-datasource'
 
 @Component({
   selector: 'app-user-management',
@@ -13,11 +14,11 @@ import { map } from 'rxjs/operators'
 export class UserManagementComponent implements OnInit {
   @ViewChild(MatPaginator) paginator: MatPaginator
   @ViewChild(MatSort) sort: MatSort
+  dataSource: UserManagementDataSource
 
-  dataSource = new MatTableDataSource([{}])
+  // dataSource = new MatTableDataSource([{}])
 
   displayedColumns: string[] = [
-    // 'position',
     // 'incognito',
     'picture',
     'id',
@@ -32,42 +33,47 @@ export class UserManagementComponent implements OnInit {
   // dataSource: MatTableDataSource<UserData>;
 
   constructor(private fss: FirestoreService) {
-    this.fss.mockUser$
-      .pipe(
-        map(ref => {
-          return ref.map((user: User) => {
-            const newUserObj: UserData = {
-              // position: 1,
-              id: user.uid,
-              picture:
-                user.photoURL || 'assets/placeholders/placeholder-user.svg',
-              name: user.displayName,
-              email: user.email,
-              campus: user.campus,
-              incognito: user.incognitoMode,
-              role: user.role,
-              points: user.points
-            }
-            return newUserObj
-          })
-        })
-      )
-      .subscribe(users => {
-        this.dataSource = new MatTableDataSource(users)
-        console.log('THIS.DATASOURCE', this.dataSource)
-      })
+    // this.fss.users$
+    //   .pipe(
+    //     map(ref => {
+    //       return ref.map((user: User) => {
+    //         const newUserObj: UserData = {
+    //           // position: 1,
+    //           id: user.uid,
+    //           picture:
+    //             user.photoURL || 'assets/placeholders/placeholder-user.svg',
+    //           name: user.displayName,
+    //           email: user.email,
+    //           campus: user.campus,
+    //           incognito: user.incognitoMode,
+    //           role: user.role,
+    //           points: user.points
+    //         }
+    //         return newUserObj
+    //       })
+    //     })
+    //   )
+    //   .subscribe(users => {
+    //     this.dataSource = new MatTableDataSource(users)
+    //     console.log('THIS.DATASOURCE', this.dataSource)
+    //   })
   }
 
   ngOnInit() {
-    this.dataSource.paginator = this.paginator
-    this.dataSource.sort = this.sort
+    this.dataSource = new UserManagementDataSource(
+      this.paginator,
+      this.sort,
+      this.fss
+    )
+    // this.dataSource.paginator = this.paginator
+    // this.dataSource.sort = this.sort
   }
 
-  applyFilter(filterValue: string) {
-    this.dataSource.filter = filterValue.trim().toLowerCase()
-
-    if (this.dataSource.paginator) {
-      this.dataSource.paginator.firstPage()
-    }
-  }
+  // applyFilter(filterValue: string) {
+  //   this.dataSource.filter = filterValue.trim().toLowerCase()
+  //
+  //   if (this.dataSource.paginator) {
+  //     this.dataSource.paginator.firstPage()
+  //   }
+  // }
 }