+++ src/app/pages/profile/profile.component.ts
@@ -10,11 +10,10 @@ import { ProgressSpinnerMode } from '@angular/material/progress-spinner'
 import { Environment } from '@environments/environment'
 import { User } from '@interfaces/user'
 import { AuthService } from '@services/auth.service'
-import { Ng2ImgToolsService } from 'ng2-img-tools'
+import { Ng2ImgMaxService } from 'ng2-img-max'
 import { Observable, of } from 'rxjs'
 import { finalize } from 'rxjs/operators'
 import { LeaveIncognitoComponent } from './dialogs/leave-incognito/leave-incognito.component'
-import { EditPictureComponent } from './edit-picture/edit-picture.component'
 
 @Component({
   selector: 'app-profile',
@@ -38,7 +37,7 @@ export class ProfileComponent {
     private dialog: MatDialog,
     private afs: AngularFirestore,
     private fb: FormBuilder,
-    private ng2ImgToolsService: Ng2ImgToolsService,
+    private ng2ImgMaxService: Ng2ImgMaxService,
     private storage: AngularFireStorage
   ) {
     this.auth.user$.subscribe(data => {
@@ -51,9 +50,8 @@ export class ProfileComponent {
     })
   }
 
-  uploadFile(event) {
-    // TODO: if image < 120x120 prompt an error: image it's too small
-    let imgCompressed
+  uploadFile(event: { target: { files: any[] } }) {
+    let imgCompressed: File
 
     const file = event.target.files[0]
     const filePath = `users/${this.user.uid}`
@@ -63,38 +61,35 @@ export class ProfileComponent {
       this.uploadPercent = of(0)
     }
 
-    this.ng2ImgToolsService
-      .resizeExactCrop([file], 128, 128)
-      .subscribe(imgResized => {
-        imgCompressed = new File([imgResized], this.user.uid)
+    this.ng2ImgMaxService.resize([file], 128, 128).subscribe(imgResized => {
+      imgCompressed = new File([imgResized], this.user.uid)
 
-        const task = this.storage.upload(filePath, imgCompressed)
+      const task = this.storage.upload(filePath, imgCompressed)
 
-        this.mode = 'determinate'
-        this.uploadPercent = task.percentageChanges()
+      this.mode = 'determinate'
+      this.uploadPercent = task.percentageChanges()
 
-        task
-          .snapshotChanges()
-          .pipe(
-            finalize(() => {
-              this.downloadURL = storageRef.getDownloadURL()
-              this.downloadURL.subscribe(ref => {
-                this.update(this.user, ref)
-                  .finally(() => {
-                    setTimeout(() => {
-                      this.uploadPercent = of(null)
-                    }, 4000)
-                  })
-                  .catch(error => error)
-              })
+      task
+        .snapshotChanges()
+        .pipe(
+          finalize(() => {
+            this.downloadURL = storageRef.getDownloadURL()
+            this.downloadURL.subscribe(ref => {
+              this.update(this.user, ref)
+                .finally(() => {
+                  setTimeout(() => {
+                    this.uploadPercent = of(null)
+                  }, 4000)
+                })
+                .catch(error => error)
             })
-          )
-          .subscribe()
-      })
+          })
+        )
+        .subscribe()
+    })
   }
 
-  async update(user: User, downloadURL) {
-    // Sets user data to firestore on login
+  async update(user: User, downloadURL: string) {
     const userRef: AngularFirestoreDocument<any> = this.afs.doc(
       `users/${user.uid}`
     )
@@ -110,15 +105,8 @@ export class ProfileComponent {
     }
   }
 
-  // Dialog Box
   openIncognitoDialog(): void {
     this.dialogRef = this.dialog.open(LeaveIncognitoComponent, {
-      autoFocus: false
-    })
-  }
-
-  openPictureDialog(): void {
-    this.dialogRef = this.dialog.open(EditPictureComponent, {
       autoFocus: false
     })
   }