+++ src/app/services/db.service.ts
@@ -23,7 +23,7 @@ export class DbService {
   //       })
   //     )
   // }
-
+  //
   // doc$(path) {
   //   return this.afs
   //     .doc(path)