+++ src/app/services/firestore.service.ts
@@ -1,33 +1,24 @@
 import { HttpClient } from '@angular/common/http'
 import { Injectable } from '@angular/core'
-import {
-  AngularFirestore,
-  AngularFirestoreCollection
-} from '@angular/fire/firestore'
+import { AngularFirestore } from '@angular/fire/firestore'
 import { Environment } from '@environments/environment'
-import { Log, NewLog } from '@interfaces/log'
+import { Log, LogReset } from '@interfaces/log'
 import { Marvel, Results } from '@interfaces/marvel'
 import { User } from '@interfaces/user'
 import { AuthService } from '@services/auth.service'
 import { Observable } from 'rxjs'
 import { map } from 'rxjs/operators'
-// import { DbService } from './db.service'
+// import { DbService } from './db.service';
 
 @Injectable()
 export class FirestoreService {
-  usersCollection: AngularFirestoreCollection<User>
-
   users$: Observable<User[]>
-  localUsers$: Observable<User[]>
-  orderedUsers$: Observable<User[]>
+  usersByPoints$: Observable<User[]>
+  usersByName$: Observable<User[]>
   logs$: Observable<Log[]>
-
-  mockUser$: Observable<User[]>
 
   position: number
   validPicture = []
-
-  logsCollectionRef = this.afs.collection('logs').ref
 
   constructor(
     // private db: DbService,
@@ -35,45 +26,31 @@ export class FirestoreService {
     private http: HttpClient,
     private auth: AuthService
   ) {
-    this.mockUser$ = afs.collection<User>('userTest').valueChanges()
-
     this.auth.user$.subscribe(userRef => {
       if (
         userRef.email === 'lyndall.benton@kent.edu.au' ||
         userRef.email === 'k170535@student.kent.edu.au'
       ) {
-        this.localUsers$ = afs
-          .collection<User>('users', res =>
-            res.where('campus', '==', 'Sydney').orderBy('points', 'desc')
-          )
+        this.usersByPoints$ = afs
+          .collection<User>('users', ref => ref.orderBy('points', 'desc'))
           .valueChanges()
-        this.orderedUsers$ = afs
-          .collection<User>('users', res =>
-            res.orderBy('campus', 'desc').orderBy('displayName', 'asc')
-          )
-          .valueChanges()
-        this.getLogs()
-      } else if (userRef.campus === 'Sydney') {
-        this.localUsers$ = afs
-          .collection<User>('users', res =>
-            res.where('campus', '==', 'Sydney').orderBy('points', 'desc')
-          )
-          .valueChanges()
-        this.orderedUsers$ = afs
-          .collection<User>('users', res =>
-            res.where('campus', '==', 'Sydney').orderBy('displayName', 'asc')
+        this.usersByName$ = afs
+          .collection<User>('users', ref =>
+            ref.orderBy('campus', 'desc').orderBy('displayName', 'asc')
           )
           .valueChanges()
         this.getLogs()
       } else {
-        this.localUsers$ = afs
-          .collection<User>('users', res =>
-            res.where('campus', '==', 'Melbourne').orderBy('points', 'desc')
+        this.usersByPoints$ = afs
+          .collection<User>('users', ref =>
+            ref.where('campus', '==', userRef.campus).orderBy('points', 'desc')
           )
           .valueChanges()
-        this.orderedUsers$ = afs
-          .collection<User>('users', res =>
-            res.where('campus', '==', 'Melbourne').orderBy('displayName', 'asc')
+        this.usersByName$ = afs
+          .collection<User>('users', ref =>
+            ref
+              .where('campus', '==', userRef.campus)
+              .orderBy('displayName', 'asc')
           )
           .valueChanges()
         this.getLogs()
@@ -91,7 +68,7 @@ export class FirestoreService {
           }
         }
 
-        this.localUsers$.subscribe((users: any) => {
+        this.usersByPoints$.subscribe((users: any) => {
           this.users$ = users
           for (const user of users) {
             if (user.incognitoMode === true) {
@@ -112,29 +89,6 @@ export class FirestoreService {
         })
       })
     })
-
-    // this.userRef = afs.collection('users').doc('0QspwDQ2f1ZtJrWSrcfqOylc8uZ2').ref
-
-    // let uRef: DocumentReference = this.afs
-    // .collection<User>('users').snapshotChanges()
-    // .pipe(
-    //       map(actions => {
-    //         console.log('ACTIONS', actions)
-    //         return actions.map(a => {
-    //           return a.payload.doc.ref
-    // })}))
-
-    // let userData = this.userRef.get().then(ref => {
-    //   console.log('REF', ref.data)
-    //   return ref.data
-    //   // console.log('REF', ref.data.call())
-    // })
-    //
-    // userData.then(res => {
-    //   console.log('RES', res.arguments)
-    // })
-
-    // console.log('TEST', userData)
   }
 
   generateNumber() {
@@ -165,11 +119,12 @@ export class FirestoreService {
   }
 
   // LOG FUNCTION
-  async addLog(refObj: NewLog) {
+  async addLog(refObj: LogReset | Log) {
     this.auth.user$.subscribe(adminData => {
-      const dataObj: NewLog = {
+      const dataObj = {
         ...refObj,
-        adminId: adminData.uid
+        adminId: adminData.uid,
+        adminName: adminData.displayName
       }
 
       try {
@@ -183,52 +138,14 @@ export class FirestoreService {
   getLogs() {
     return (this.logs$ = this.afs
       .collection<Log>('logs', res => res.orderBy('date', 'desc'))
-      .valueChanges()).pipe(
-      map(res => {
-        return res.map((logItem: any) => {
-          // console.log('LOGITEM', logItem)
-
-          // this.db.doc$(`/users/${logItem.userId}`).subscribe(userData => {
-          //   // console.log('USERDATA', userData)
-          //   // this.userData = userData
-          // })
-          //
-          // this.db.doc$(`/users/${logItem.adminId}`).subscribe(adminData => {
-          //   // console.log('ADMINDATA', adminData)
-          //   // this.adminData = adminData
-          // })
-
-          const newData: any = {
-            ...logItem
-          }
-          return newData
-        })
-      })
-    )
-  }
-
-  getLogsValue() {
-    return this.afs
-      .collection<Log>('logs', res => res.orderBy('date', 'desc'))
       .snapshotChanges()
       .pipe(
         map(actions => {
           return actions.map(a => {
-            const data: object = a.payload.doc.data()
-            // const id = a.payload.doc.id
-            // const additionalData = {
-            //   userId: 'string',
-            //   userPicture: 'assets/placeholders/placeholder-user',
-            //   userName: 'string',
-            //   userEmail: 'string',
-            //   userCampus: 'string',
-            //   pointsCurrent: 123,
-            //   message: 'string'
-            // }
+            const data: Log = a.payload.doc.data()
             return data
-            // return { id, ...additionalData, ...data }
           })
         })
-      )
+      ))
   }
 }